requirements.txt
numpy==1.23.5
opencv_python==4.9.0.80
Pillow==9.5.0
Pillow==10.4.0
PyYAML==6.0.1
torch==2.0.1
ultralytics==8.2.49
